+++ Framework/Source/API/ComputeContext.cpp
@@ -95,8 +95,11 @@ namespace Falcor
         {
             logWarning("ComputeContext::prepareForDispatch() - applying ComputeVars failed, most likely because we ran out of descriptors. Flushing the GPU and retrying");
             flush(true);
-            bool b = mpComputeVars->apply(const_cast<ComputeContext*>(this), mBindComputeRootSig);
-            assert(b);
+            if (!mpComputeVars->apply(const_cast<ComputeContext*>(this), mBindComputeRootSig))
+            {
+                logError("ComputeVars::applyComputeVars() - applying ComputeVars failed, most likely because we ran out of descriptors", true);
+                assert(false);
+            }
         }
     }