+++ Framework/Source/Data/HostDeviceSharedMacros.h
@@ -145,6 +145,9 @@
 #define LightPoint                  0    ///< Point light source, can be a spot light if its opening angle is < 2pi
 #define LightDirectional            1    ///< Directional light source
 #define LightArea                   2    ///< Area light source, potentially with arbitrary geometry
+#define LightAreaRect               3    ///< Quad shaped area light source
+#define LightAreaSphere             4    ///< Spherical area light source
+#define LightAreaDisc               5    ///< Disc shaped area light source
 
 #define MAX_LIGHT_SOURCES 16