+++ Framework/Source/Effects/ToneMapping/ToneMapping.h
@@ -99,6 +99,26 @@ namespace Falcor
         */
         void setWhiteScale(float whiteScale);
 
+        /** Get the tonemapping operator type.
+		*/
+        Operator getOperator() const { return mOperator; }
+        
+        /** Get tonemapper exposure key value.
+		*/
+        float getExposureKey() const { return mConstBufferData.exposureKey; }
+        
+        /** Gets the maximal luminance to be consider as pure white. 
+		*/
+        float getWhiteMaxLuminance() const { return mConstBufferData.whiteMaxLuminance; }
+        
+        /** Gets the luminance texture LOD to use when fetching average luminance values. 
+		*/
+        float getLuminanceLod() const { return mConstBufferData.luminanceLod; }
+        
+        /** Gets the white-scale used in Uncharted 2 tone mapping. 
+		*/
+        float getWhiteScale() const { return mConstBufferData.whiteScale; }
+
     private:
         ToneMapping(Operator op);
         void createLuminanceFbo(Fbo::SharedPtr pSrcFbo);