+++ Framework/Source/Graphics/Scene/Scene.cpp
@@ -85,10 +85,10 @@ namespace Falcor
                         sceneAABB = BoundingBox::fromUnion(sceneAABB, pInst->getBoundingBox());
                     }
                 }
-
-                mCenter = sceneAABB.center;
-                mRadius = length(sceneAABB.extent);
             }
+            
+            mCenter = sceneAABB.center;
+            mRadius = length(sceneAABB.extent) * 0.5f;
 
             // Update light extents
             for (auto& pLight : mpLights)
@@ -333,6 +333,7 @@ namespace Falcor
             const Model::SharedPtr& pModel = getModel(modelId);
             if (pModel)
             {
+                // TODO: Create area lights per model instance
                 std::vector<AreaLight::SharedPtr> areaLights = createAreaLightsForModel(pModel.get());
                 mpAreaLights.insert(mpAreaLights.end(), areaLights.begin(), areaLights.end());
             }